Improved on the heap safety when creating events
Moved the formatting of payloads to the `create_event` method to allow
us to perform additional sanity checks. This avoids possible heap
pollution. With the help of some additional inline functions the calling
pattern for event creation is simplified a lot.

Also fixed a couple of warnings that Clang found wrt potential use of
undefined variables.