Implemented support for OUI lookups
When a OUI vendor listing is supplied through the configuration, it will
be used to lookup vendor names when issuing events. When *no* vendor
listing is supplied, no lookups will be performed. The code is inspired
on how Wireshark loads and uses vendor lookups.

When receiving SIGUSR2, the configured OUI list is reloaded from disk.

Closes #6.